Refactor TablatureEditorPlus. It is now a new component, Editor, and is better composed to improve readability and seperation of concerns. The Bar component has also been improved
